requirements.txt
streamlit 
langchain 
google-generativeai
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv